+++ src/domain/order/Order.java
@@ -10,7 +10,8 @@ public class Order {
     private String requestMessage;          // 3 요청사항
     private Instant orderDateTime;          // 4 주문 일시
     private int waitingNumber = 0;           // 5 대기번호
-//====================================================생성자
+
+    //====================================================생성자
     public Order(double totalPrice, Cart cart, String requestMessage, Instant orderDateTime, int waitingNumber) {
         this.totalPrice = cart.getTotalPrice();
         this.cart = cart;
@@ -18,7 +19,8 @@ public class Order {
         this.orderDateTime = orderDateTime;
         this.waitingNumber = waitingNumber;
     }
-//=================================================메서드 구역
+
+    //=================================================메서드 구역
     public Cart getCart() {
         return cart;
     }                                       // 2 장바구니
@@ -34,6 +36,10 @@ public class Order {
     public int getWaitingNumber() {
         return waitingNumber;
     }                                       //5 대기번호
+
+    public double getTotalPrice() {
+        return totalPrice;
+    }
 
     public void inputRequestMessage(String message) throws IllegalArgumentException {
         if (message.length() > 20) {
